Rework expires_on logic for managed identity
