+++ sdk/identity/identity/src/credentials/managedIdentityCredential/models.ts
@@ -2,12 +2,7 @@
 // Licensed under the MIT license.
 
 import { AccessToken, GetTokenOptions } from "@azure/core-auth";
-import { IdentityClient, TokenResponseParsedBody } from "../../client/identityClient";
-
-/**
- * @internal
- */
-export type MSIExpiresInParser = (requestBody: TokenResponseParsedBody) => number;
+import { IdentityClient } from "../../client/identityClient";
 
 /**
  * @internal
